Make booth_conf->site_bits 64bit, same as ticket_conf->proposal_acknowledges.

+++ src/config.h
@@ -81,7 +81,7 @@ struct booth_config {
     uint16_t port;
 
 	/** Stores the OR of the individual host bitmasks. */
-	uint32_t site_bits;
+	uint64_t site_bits;
 
     int site_count;
     struct booth_site site[MAX_NODES];
